Add feature to call assert(...).(not)equal with sequences.
Sequences (arrays etc) don't conform to the Equatable protocol and
therefore couldn't be used in an assert call. For example take a
ViewModel that produces lists of objects to hook up to a TableView. The
only option was to check .is(Not)Empty and not assert the contents of
the sequences.